Make it pip installable
