+++ setup.cfg
@@ -1,0 +1,23 @@
+[metadata]
+name = streamlit-multipage
+version = 0.0.8
+author = Yan Almeida, Ezequiel Leonardo Castaño
+author_email = author@example.com
+description = A small package that allows you to create multi page streamlit apps
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/ELC/streamlit-multipage
+project_urls =
+    Bug Tracker = https://github.com/ELC/streamlit-multipage/issues
+classifiers =
+    Programming Language :: Python :: 3
+    Operating System :: OS Independent
+
+[options]
+package_dir =
+    = src
+packages = find:
+python_requires = >=3.6
+
+[options.packages.find]
+where = src
